Move 'dto' package out of 'controllers'. Some little changes
