+++ src/main/java/smida/techtask/controllers/ReportController.java
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
-import smida.techtask.controllers.dto.ReportAndCompanyIdDto;
+import smida.techtask.dto.ReportAndCompanyIdDto;
 import smida.techtask.services.ReportService;
 
 import java.util.List;
