+++ src/main/java/smida/techtask/controllers/handlers/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package smida.techtask.controllers;
+package smida.techtask.controllers.handlers;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -6,16 +6,16 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.context.request.WebRequest;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
-import smida.techtask.annotations.CompanyNotFoundException;
-import smida.techtask.controllers.dto.ErrorDto;
+import smida.techtask.dto.security.ErrorDto;
+import smida.techtask.exceptions.BaseNotFoundException;
 
 import java.time.LocalDateTime;
 
 @RestControllerAdvice
 public class GlobalExceptionHandler extends ResponseEntityExceptionHandler {
 
-    @ExceptionHandler(value = CompanyNotFoundException.class)
-    public ResponseEntity<ErrorDto> missingRequestHeader(CompanyNotFoundException exception, WebRequest webRequest) {
+    @ExceptionHandler(value = BaseNotFoundException.class)
+    public ResponseEntity<ErrorDto> missingRequestHeader(BaseNotFoundException exception, WebRequest webRequest) {
         ErrorDto errorDetails = new ErrorDto(
                 LocalDateTime.now(),
                 exception.getMessage(),
