+++ src/main/java/smida/techtask/dto/CompanyDto.java
@@ -1,4 +1,4 @@
-package smida.techtask.controllers.dto;
+package smida.techtask.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
