+++ src/main/java/smida/techtask/dto/ReportAndCompanyIdDto.java
@@ -1,6 +1,5 @@
-package smida.techtask.controllers.dto;
+package smida.techtask.dto;
 
-import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 
 import java.math.BigDecimal;
@@ -8,15 +7,15 @@ import java.time.Instant;
 import java.util.UUID;
 
 @Data
-public class ReportDto {
+public class ReportAndCompanyIdDto {
 
     private UUID id;
 
     private Instant reportDate;
 
-    @NotNull
     private BigDecimal totalRevenue;
 
-    @NotNull
     private BigDecimal netProfit;
+
+    private UUID companyId;
 }
