+++ src/main/java/smida/techtask/dto/ReportDto.java
@@ -1,5 +1,6 @@
-package smida.techtask.controllers.dto;
+package smida.techtask.dto;
 
+import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 
 import java.math.BigDecimal;
@@ -7,15 +8,15 @@ import java.time.Instant;
 import java.util.UUID;
 
 @Data
-public class ReportAndCompanyIdDto {
+public class ReportDto {
 
     private UUID id;
 
     private Instant reportDate;
 
+    @NotNull
     private BigDecimal totalRevenue;
 
+    @NotNull
     private BigDecimal netProfit;
-
-    private UUID companyId;
 }
