+++ src/main/java/smida/techtask/dto/security/LoginDto.java
@@ -1,4 +1,4 @@
-package smida.techtask.controllers.dto;
+package smida.techtask.dto.security;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
