+++ src/main/java/smida/techtask/dto/security/RegistrationDto.java
@@ -1,4 +1,4 @@
-package smida.techtask.controllers.dto;
+package smida.techtask.dto.security;
 
 
 import jakarta.validation.constraints.NotBlank;
