+++ src/main/java/smida/techtask/entities/Company.java
@@ -36,7 +36,7 @@ public class Company {
     @CreatedDate
     private Instant createdAt;
 
-    @OneToMany(mappedBy = "company", cascade = CascadeType.ALL, orphanRemoval = true, fetch = FetchType.EAGER)
+    @OneToMany(mappedBy = "company", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<Report> reports = new ArrayList<>();
 
 }
