+++ src/main/java/smida/techtask/entities/ReportDetails.java
@@ -6,7 +6,7 @@ import org.springframework.data.mongodb.core.mapping.Document;
 
 import java.util.UUID;
 
-@Document
+@Document("report_details")
 @AllArgsConstructor
 @NoArgsConstructor
 @Setter
