+++ src/main/java/smida/techtask/mappers/CompanyReportsMapper.java
@@ -3,8 +3,8 @@ package smida.techtask.mappers;
 import org.mapstruct.Mapper;
 import org.mapstruct.MappingConstants;
 import org.mapstruct.ReportingPolicy;
-import smida.techtask.controllers.dto.CompanyReportsDto;
-import smida.techtask.controllers.dto.ReportDto;
+import smida.techtask.dto.CompanyReportsDto;
+import smida.techtask.dto.ReportDto;
 import smida.techtask.entities.Company;
 import smida.techtask.entities.Report;
