+++ src/main/java/smida/techtask/repositories/ReportDetailsRepository.java
@@ -1,4 +1,10 @@
 package smida.techtask.repositories;
 
-public interface ReportDetailsRepository {
+import org.springframework.data.mongodb.repository.MongoRepository;
+import smida.techtask.entities.ReportDetails;
+
+import java.util.UUID;
+
+public interface ReportDetailsRepository extends MongoRepository<ReportDetails, UUID> {
+
 }
