+++ src/main/java/smida/techtask/services/impl/ReportServiceImpl.java
@@ -2,7 +2,7 @@ package smida.techtask.services.impl;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import smida.techtask.controllers.dto.ReportAndCompanyIdDto;
+import smida.techtask.dto.ReportAndCompanyIdDto;
 import smida.techtask.repositories.managers.ReportManager;
 import smida.techtask.mappers.ReportAndCompanyIdMapper;
 import smida.techtask.services.ReportService;
